Add files via upload
BSM_main utilizes the Black Scholes Model class to read dataframes of option pricing data, displaying many theoretical plots
BSM stores the class
utils stores a bunch of functions needed to both parse the yahoo finance api data and manipulate it with pandas and numpy
neural_network is a model network created to learn BSM, and does a really good job. Might be trivial as the basic BSM is deterministic.